Add migration vertion with random inserts
