+++ GiftCertificatesSystem/web/src/main/resources/application.properties
@@ -19,6 +19,7 @@ spring.security.oauth2.client.registration.github.client-secret=${github.client-
 # FLYWAY
 spring.flyway.enabled=true
 spring.flyway.baseline-on-migrate=true
+spring.flyway.baseline-version=2
 
 # SPRING-BOOT-CLIENT-APP
 spring.boot.admin.client.url=http://localhost:8080
